[Fixed #1] Added test to check if system logging is working.
The used journalctl command produces only 2 lines if there are logs for the current
date:
1. Logging begin and end summary
2. The most recent log for the current system date

else it produces only the first line, i.e., logging begin and end summary.
The concise output incurs very minimal data transfer from the
server to the fabric client and less data to work with for assertion.